Move `deepClone` to shared utils
