+++ src/adapter/protocol/legacy/operationsV1.ts
@@ -1,7 +1,7 @@
 import { MsgTypes } from "../events";
 import { parseTable } from "../string-table";
 import { Store } from "../../../view/store/types";
-import { deepClone } from "../../../view/components/profiler/flamegraph/modes/adjustNodesToRight";
+import { deepClone } from "../../shared/utils";
 
 export function applyOperationsV1(store: Store, data: number[]) {
 	const nodes = new Map(store.nodes.$);
