+++ src/adapter/protocol/operations.ts
@@ -1,7 +1,7 @@
 import { ID, Tree } from "../../view/store/types";
 import { parseTable } from "./string-table";
 import { MsgTypes } from "./events";
-import { deepClone } from "../../view/components/profiler/flamegraph/modes/adjustNodesToRight";
+import { deepClone } from "../shared/utils";
 import { RenderReasonMap } from "../shared/renderReasons";
 import { ParsedStats, parseStats } from "../shared/stats";
