+++ src/adapter/shared/utils.ts
@@ -15,3 +15,8 @@ export function traverse<T extends SharedVNode>(
 		}
 	}
 }
+
+// eslint-disable-next-line @typescript-eslint/ban-types
+export function deepClone<T extends object>(obj: T): T {
+	return JSON.parse(JSON.stringify(obj));
+}
